+++ internal/compiler/lookup.rs
@@ -802,6 +802,7 @@ impl LookupObject for MathFunctions {
             .or_else(|| f("asin", BuiltinFunctionReference(BuiltinFunction::ASin, sl())))
             .or_else(|| f("acos", BuiltinFunctionReference(BuiltinFunction::ACos, sl())))
             .or_else(|| f("atan", BuiltinFunctionReference(BuiltinFunction::ATan, sl())))
+            .or_else(|| f("atan2", BuiltinFunctionReference(BuiltinFunction::ATan2, sl())))
             .or_else(|| f("log", BuiltinFunctionReference(BuiltinFunction::Log, sl())))
             .or_else(|| f("pow", BuiltinFunctionReference(BuiltinFunction::Pow, sl())))
     }