+++ CHANGELOG.md
@@ -16,6 +16,10 @@ All notable changes to this project are documented in this file.
  - Do not trigger `current-item-changed` on `StandardListView` if `current-item` is set on the same value.
  - Fixed `TimePickerPopup` does not open minute view by click on selected hour.
 
+### Slint Language
+
+ - Added math function `atan2`.
+
 ## [1.7.2] - 2024-08-14
 
 ### General